# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat APP API",
        "description": "API documentation for Chat Application - A social media chat application with posts, comments, likes, friendships, and real-time messaging",
        "contact": {
            "email": "support@chatapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/conversations": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "summary": "Get all conversations",
                "description": "Get all conversations for the authenticated user with last message and unread count",
                "operationId": "5e013ca2cc574b38237e805084bc9a00",
                "responses": {
                    "200": {
                        "description": "List of conversations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Jane Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jane@example.com"
                                            },
                                            "profile_image": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "lastMessage": {
                                                "type": "string",
                                                "example": "Hello!",
                                                "nullable": true
                                            },
                                            "lastMessageAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            },
                                            "unreadCount": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/conversations/friends": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "summary": "Get all friends",
                "description": "Get all accepted friends for the authenticated user",
                "operationId": "fc7d13d514f3fed1db07f4fd3efa9049",
                "responses": {
                    "200": {
                        "description": "List of friends",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Jane Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jane@example.com"
                                            },
                                            "profile_image": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "last_seen_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/conversations/user/{userId}": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "summary": "Get or create conversation with user",
                "description": "Get or create a conversation with a specific friend",
                "operationId": "d5e8b11ad2cffb1adaae7eb46145be89",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID to get/create conversation with",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation and messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "conversation": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jane Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "jane@example.com"
                                                },
                                                "profile_image": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    },
                                                    "sender_id": {
                                                        "type": "integer"
                                                    },
                                                    "sender_name": {
                                                        "type": "string"
                                                    },
                                                    "seen": {
                                                        "type": "boolean"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Friendship not found or not accepted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/conversations/{conversation}/mark-as-seen": {
            "post": {
                "tags": [
                    "Conversations"
                ],
                "summary": "Mark messages as seen",
                "description": "Mark all unread messages in a conversation as seen",
                "operationId": "c553bf0d961576a2223bfc998f7d5973",
                "parameters": [
                    {
                        "name": "conversation",
                        "in": "path",
                        "description": "Conversation ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages marked as seen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "marked_count": {
                                            "description": "Number of messages marked as seen",
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - user not part of conversation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication using Laravel session-based authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}